requirements.txt
Flask
Flask-SQLAlchemy
# SQLAlchemy
psycopg2-binary

# Huso horario
# backports.zoneinfo # laburo
tzdata # casa


flask-wtf